Separate useChatList hook and constants, types
