+++ app/chat/types.ts
@@ -1,0 +1,13 @@
+export interface Badge {
+    name: string;
+    src: string;
+}
+
+export interface Chat {
+    uuid: string;
+    nickname: string;
+    badges: Badge[];
+    color: number | string;
+    emojis: Record<string, string>;
+    message: string;
+}
